Fixes to the merge Add a "Enter code manually" button on ScanScreen (WIP)

+++ app/src/main/java/ca/uqac/etu/jcid/chadal/data/ShoppingListEntity.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "shopping_list")
 data class ShoppingListEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val budget: Int,
+    val budget: Double,
     val date: String,
     val article: Int
